added esc keybind as alias for exit
because esc is a prefix for existing useful binds, this required changing the
way input is taken. Whereas previously, input was aken one char at a a time,
now it takes as many chars as it can at once. A keybind will only match if it
matches the whole input (so esc will match "\x1b", but not "\x1b[").